+++ src/v7400/data/mesh.rs
@@ -1,10 +1,16 @@
 //! Mesh data.
 
 pub use self::{
-    control_point::{ControlPointIndex, ControlPoints},
-    polygon_vertex_index::{PolygonIndex, PolygonVertex, PolygonVertexIndex, PolygonVertices},
-    triangle_vertex_index::{TriangleIndex, TriangleVertexIndex, TriangleVertices},
+    control_point::ControlPointIndex,
+    polygon_vertex_index::{
+        IntoCpiWithPolyVerts, PolygonIndex, PolygonVertex, PolygonVertexIndex, PolygonVertices,
+    },
+    triangle_vertex_index::{
+        IntoCpiWithTriVerts, IntoPvWithTriVerts, TriangleIndex, TriangleVertexIndex,
+        TriangleVertices,
+    },
 };
+pub(crate) use self::{control_point::ControlPoints, polygon_vertex_index::RawPolygonVertices};
 
 mod control_point;
 pub mod layer;