+++ src/v7400/data/mesh/control_point.rs
@@ -1,7 +1,5 @@
 //! Control point.
 
-use crate::v7400::data::mesh::PolygonVertex;
-
 /// Control point index (in other words, polygon vertex).
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct ControlPointIndex(u32);
@@ -22,16 +20,11 @@ impl ControlPointIndex {
     pub fn get_u32(self) -> u32 {
         self.to_u32()
     }
-
-    /// Creates a new `ControlPointIndex` from the given `PolygonVertex`.
-    pub(crate) fn from_pv(pv: PolygonVertex) -> Self {
-        Self(pv.to_u32())
-    }
 }
 
 /// Control points.
 #[derive(Debug, Clone, Copy)]
-pub struct ControlPoints<'a> {
+pub(crate) struct ControlPoints<'a> {
     /// Control points.
     data: &'a [f64],
 }
@@ -43,38 +36,11 @@ impl<'a> ControlPoints<'a> {
     }
 
     /// Returns a control point at the given index.
-    pub fn get_cp_f64(&self, index: ControlPointIndex) -> Option<[f64; 3]> {
+    pub(crate) fn get(&self, index: ControlPointIndex) -> Option<[f64; 3]> {
         let i3 = index.to_u32() as usize * 3;
         if self.data.len() < i3 + 2 {
             return None;
         }
         Some([self.data[i3], self.data[i3 + 1], self.data[i3 + 2]])
-    }
-
-    /// Returns a control point at the given index.
-    pub fn get_cp_f32(&self, index: ControlPointIndex) -> Option<[f32; 3]> {
-        let i3 = index.to_u32() as usize * 3;
-        if self.data.len() < i3 + 2 {
-            return None;
-        }
-        Some([
-            self.data[i3] as f32,
-            self.data[i3 + 1] as f32,
-            self.data[i3 + 2] as f32,
-        ])
-    }
-
-    /// Returns iterator of `[f32; 3]` vertices.
-    pub fn iter_f32(&self) -> impl Iterator<Item = [f32; 3]> + 'a {
-        self.data
-            .chunks_exact(3)
-            .map(|arr| [arr[0] as f32, arr[1] as f32, arr[2] as f32])
-    }
-
-    /// Returns iterator of `[f64; 3]` vertices.
-    pub fn iter_f64(&self) -> impl Iterator<Item = [f64; 3]> + 'a {
-        self.data
-            .chunks_exact(3)
-            .map(|arr| [arr[0], arr[1], arr[2]])
     }
 }