+++ src/v7400/data/mesh/layer/common.rs
@@ -237,7 +237,7 @@ impl LayerContentIndex {
     }
 
     /// Returns the layer content index for the corresponding control point.
-    pub(crate) fn control_ponint_data_from_triangle_vertices(
+    pub(crate) fn control_point_data_from_triangle_vertices(
         reference_info: ReferenceInformation<'_>,
         mapping_mode: MappingMode,
         triangle_vertices: &TriangleVertices<'_>,
@@ -247,20 +247,24 @@ impl LayerContentIndex {
         let index = match mapping_mode {
             MappingMode::None | MappingMode::ByEdge => bail!("Unsupported mapping mode: {:?}"),
             MappingMode::ByControlPoint => {
-                let cpi = triangle_vertices.get_control_point(tri_vi).ok_or_else(|| {
-                    format_err!("Failed to get control point index: tri_vi={:?}", tri_vi)
-                })?;
+                let cpi = triangle_vertices
+                    .control_point_index(tri_vi)
+                    .ok_or_else(|| {
+                        format_err!("Failed to get control point index: tri_vi={:?}", tri_vi)
+                    })?;
                 reference_info.get_direct(cpi.to_u32() as usize)?
             }
             MappingMode::ByPolygonVertex => {
-                let pvi = triangle_vertices.get_pvi(tri_vi).ok_or_else(|| {
-                    format_err!("Failed to get polygon vertex index: tri_vi={:?}", tri_vi)
-                })?;
+                let pvi = triangle_vertices
+                    .polygon_vertex_index(tri_vi)
+                    .ok_or_else(|| {
+                        format_err!("Failed to get polygon vertex index: tri_vi={:?}", tri_vi)
+                    })?;
                 reference_info.get_direct(pvi.to_usize() as usize)?
             }
             MappingMode::ByPolygon => {
                 let poly_i = triangle_vertices
-                    .get_polygon_index(tri_vi.triangle_index())
+                    .polygon_index(tri_vi.triangle_index())
                     .ok_or_else(|| {
                         format_err!("Failed to get polygon vertex index: tri_vi={:?}", tri_vi)
                     })?;