+++ src/v7400/object/geometry/mesh.rs
@@ -3,7 +3,7 @@
 use failure::{format_err, Error};
 
 use crate::v7400::{
-    data::mesh::{layer::LayerHandle, ControlPoints, PolygonVertices},
+    data::mesh::{layer::LayerHandle, ControlPoints, PolygonVertices, RawPolygonVertices},
     object::{deformer, geometry::GeometryHandle, model, TypedObjectHandle},
 };
 
@@ -49,7 +49,7 @@ impl<'a> MeshHandle<'a> {
     }
 
     /// Returns control points.
-    pub fn control_points(&self) -> Result<ControlPoints<'a>, Error> {
+    pub(crate) fn control_points(&self) -> Result<ControlPoints<'a>, Error> {
         self.node()
             .children_by_name("Vertices")
             .next()
@@ -67,8 +67,8 @@ impl<'a> MeshHandle<'a> {
             })
     }
 
-    /// Returns polygon vertex indices.
-    pub fn polygon_vertex_indices(&self) -> Result<PolygonVertices<'a>, Error> {
+    /// Returns polygon vertices without control points.
+    pub(crate) fn raw_polygon_vertices(&self) -> Result<RawPolygonVertices<'a>, Error> {
         self.node()
             .children_by_name("PolygonVertexIndex")
             .next()
@@ -79,13 +79,20 @@ impl<'a> MeshHandle<'a> {
             .get(0)
             .ok_or_else(|| format_err!("`PolygonVertexIndex` node has no attributes"))?
             .get_arr_i32_or_type()
-            .map(PolygonVertices::new)
+            .map(RawPolygonVertices::new)
             .map_err(|ty| {
                 format_err!(
                     "`PolygonVertexIndex` has wrong type attribute: expected `[i32]` but got {:?}`",
                     ty
                 )
             })
+    }
+
+    /// Returns polygon vertices (control point indices) and control points.
+    pub fn polygon_vertices(&self) -> Result<PolygonVertices<'a>, Error> {
+        let control_points = self.control_points()?;
+        let raw_polygon_vertices = self.raw_polygon_vertices()?;
+        Ok(PolygonVertices::new(control_points, raw_polygon_vertices))
     }
 
     /// Returns layers.